+++ Makefile
@@ -44,7 +44,7 @@ OFILES = assembled_chunk.o \
 	udp_packet_list.o \
 	udp_packet_ringbuf.o
 
-INCFILES=ch_frb_io.hpp ch_frb_io_internals.hpp ringbuf.hpp ringbuf.cpp
+INCFILES=ch_frb_io.hpp ch_frb_io_internals.hpp ringbuf.hpp ringbuf-impl.hpp
 LIBFILES=libch_frb_io.so
 INSTALLED_BINARIES=ch-show-intensity-file
 INSTALLED_SCRIPTS=ch-plot-intensity-file
