+++ ringbuf-impl.hpp
